+++ src/main/java/solutions/thex/reporeporter/svg/resolver/badge/IconResolver.java
@@ -1,30 +1,30 @@
 package solutions.thex.reporeporter.svg.resolver.badge;
 
 import solutions.thex.reporeporter.svg.SvgResolver;
-import solutions.thex.reporeporter.svg.generator.badge.LogoGenerator;
+import solutions.thex.reporeporter.svg.generator.badge.IconGenerator;
 
 import java.io.IOException;
 import java.util.Map;
 
 /**
  * An implementation of {@link solutions.thex.reporeporter.svg.SvgResolver} which resolves given parameters by user to
- * a {@link solutions.thex.reporeporter.svg.generator.badge.LogoGenerator} instance.
+ * a {@link solutions.thex.reporeporter.svg.generator.badge.IconGenerator} instance.
  *
  * @author Soroush Shemshadi
  * @version 1.0.0
  * @since 1.0.0
  */
-public class LogoResolver extends SvgResolver {
+public class IconResolver extends SvgResolver {
 
     @Override
     public String resolve(Map<String, String> params) throws IOException {
         final String bg = resolveBG(params.get("bg"));
 
-        return LogoGenerator.builder()//
+        return IconGenerator.builder()//
                 .theme(params.get("theme"))//
                 .size(params.get("size"))//
-                .title(params.get("logo"))//
-                .logo(resolveLogo(params.get("logo"), resolveColor(params.get("color"), bg)))//
+                .title(params.get("icon"))//
+                .icon(resolveIcon(params.get("icon"), resolveColor(params.get("color"), bg)))//
                 .link(params.get("link"))//
                 .width(resolveWidth(params.get("size"), params.get("title")))//
                 .height(resolveHeight(params.get("size")))//