+++ src/main/java/solutions/thex/reporeporter/svg/responseWrapper/badge/IconAsResponseWrapper.java
@@ -4,30 +4,30 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import solutions.thex.reporeporter.controller.error.ErrorAsBadge;
 import solutions.thex.reporeporter.svg.SvgAsResponseWrapper;
-import solutions.thex.reporeporter.svg.resolver.badge.LogoResolver;
+import solutions.thex.reporeporter.svg.resolver.badge.IconResolver;
 
 import java.io.IOException;
 import java.util.Map;
 
 /**
  * An implementation of {@link solutions.thex.reporeporter.svg.SvgAsResponseWrapper} which wrap up generated
- * logo badge SVG as a response.
+ * icon badge SVG as a response.
  *
  * @author Soroush Shemshadi
  * @version 1.0.0
  * @since 1.0.0
  */
-public class LogoAsResponseWrapper implements SvgAsResponseWrapper {
+public class IconAsResponseWrapper implements SvgAsResponseWrapper {
 
     @Override
     public ResponseEntity<String> wrap(Map<String, String> params) throws IOException {
-        if ("-1".equals(params.get("logo")))
+        if ("-1".equals(params.get("icon")))
             return new ResponseEntity<>(//
-                    new ErrorAsBadge(422, "Title or logo not provided!").toString(),//
+                    new ErrorAsBadge(422, "Title or icon not provided!").toString(),//
                     HttpStatus.UNPROCESSABLE_ENTITY);
 
         return new ResponseEntity<>(//
-                new LogoResolver().resolve(params) //
+                new IconResolver().resolve(params) //
                 , HttpStatus.OK);
     }