+++ src/main/java/solutions/thex/reporeporter/svg/responseWrapper/badge/LinkAsResponseWrapper.java
@@ -21,9 +21,9 @@ import java.util.Map;
 public class LinkAsResponseWrapper implements SvgAsResponseWrapper {
 
     public ResponseEntity<String> wrap(Map<String, String> params) throws IOException {
-        if ("-1".equals(params.get("title")) || "-1".equals(params.get("logo")))
+        if ("-1".equals(params.get("title")) || "-1".equals(params.get("icon")))
             return new ResponseEntity<>(//
-                    new ErrorAsBadge(422, "Title or logo not provided!").toString(),//
+                    new ErrorAsBadge(422, "Title or icon not provided!").toString(),//
                     HttpStatus.UNPROCESSABLE_ENTITY);
 
         return new ResponseEntity<>(//