+++ src/main/java/solutions/thex/reporeporter/svg/responseWrapper/badge/ProfileAsResponseWrapper.java
@@ -22,7 +22,7 @@ public class ProfileAsResponseWrapper implements SvgAsResponseWrapper {
     public ResponseEntity<String> wrap(Map<String, String> params) throws IOException {
         if ("-1".equals(params.get("id")))
             return new ResponseEntity<>(//
-                    new ErrorAsBadge(422, "Title or logo not provided!").toString(),//
+                    new ErrorAsBadge(422, "Title or icon not provided!").toString(),//
                     HttpStatus.UNPROCESSABLE_ENTITY);
 
         return new ResponseEntity<>(//